Refactor: 타입 폴더 바깥으로 합침
