+++ next.config.mjs
@@ -7,7 +7,7 @@ const __dirname = path.dirname(__filename);
 
 const nextConfig = {
   sassOptions: {
-    includePaths: [path.join(__dirname, "src/app/styles")],
+    includePaths: [path.join(__dirname, "src/styles")],
     additionalData: `
     @import "_variables";
     @import "_mixins";`,
