+++ src/lib/api/auth.ts
@@ -1,7 +1,7 @@
 "use server";
 
 import { cookies } from "next/headers";
-import { PostAuth } from "@/app/types/auth/authTypes";
+import { PostAuth } from "@/types/auth/authTypes";
 import API_URL from "@/constants/config";
 
 const setCookie = (name: string, value: string) => {
